Created CRUD Task and task filtering

+++ app/Http/Controllers/LabelController.php
@@ -50,6 +50,12 @@ class LabelController extends Controller
 
     public function destroy(Label $label): RedirectResponse
     {
+        if ($label->tasks()->exists()) {
+            flash()->error(__('label.has_tasks'));
+            return back()->withErrors([
+                'message' => __('label.has_tasks'),
+            ]);
+        }
         $label->delete();
         flash()->success(__('label.deleted'));
         return redirect(route('labels.index'));
